+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix: The false alarm that Elite Bionics Framework will detect itself violating the EBF protocol.
+
 ## 1.0.23  (2026-05-18)
 
 * Compatibility: Pick Up And Haul compatibility patch will recognize the unofficial fork (package id: mehni.pickupandhaul.unoffical), though this version is obsolete.